# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ICTU AI Judge API
  description: RESTful API spec for ICTU AI Judge Platform (FastAPI Backend)
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server

# Bảo mật: Yêu cầu JWT Token truyền ngầm qua HttpOnly Cookie thay vì Header Bearer.
security:
  - cookieAuth: []
paths:
...
  /challenges/{id}/submissions:
    get:
      summary: Lịch sử nộp bài của Đội
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Danh sách lượt nộp (Phân trang).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Submission'

    post:
      summary: Upload file Nộp bài
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File CSV dự đoán
      responses:
        '201':
          description: File hợp lệ, bắt đầu chấm điểm ngầm (Trạng thái PENDING).
        '429':
          description: Vi phạm Rate Limit (Xin chờ X phút).
        '409':
          description: Trùng Hash MD5 (Spam, file này vừa nộp rồi).
...
  /admin/users:
    get:
      summary: Danh sách Sinh viên (Admin)
      tags:
        - Admin
      parameters:
        - name: q
          in: query
          description: Tìm kiếm theo email hoặc họ tên
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Danh sách user (phân trang).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserResponse'
        '403':
          description: Không có quyền Admin.
  /admin/users/{id}:
    patch:
      summary: Khóa / Mở khóa tài khoản Sinh viên (Admin)
      tags:
        - Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
                  description: true = mở khóa, false = khóa tài khoản
      responses:
        '200':
          description: Cập nhật trạng thái tài khoản thành công.
        '403':
          description: Không có quyền Admin.
  /admin/challenges/test-metric:
    post:
      summary: Chạy thử Metric (Admin)
      tags:
        - Admin
      description: Chạy thử file Ground Truth, Submission và Metric Script trong Sandbox để kiểm tra.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - ground_truth
                - submission
                - metric_name
              properties:
                ground_truth:
                  type: string
                  format: binary
                  description: File đáp án mẫu (CSV)
                submission:
                  type: string
                  format: binary
                  description: File bài nộp mẫu (CSV)
                metric_script:
                  type: string
                  format: binary
                  description: (Tùy chọn) Script python chấm điểm thủ công
                metric_name:
                  type: string
                  description: Tên metric (ví dụ ACCURACY, F1_SCORE, CUSTOM)
      responses:
        '200':
          description: Chạy Sandbox thành công, trả về điểm.
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: number
        '400':
          description: Lỗi Sandbox hoặc lỗi file, trả về chi tiết.
        '403':
          description: Không có quyền Admin.
  /admin/challenges/{challenge_id}/submissions:
    get:
      summary: Lấy danh sách tất cả bài nộp của cuộc thi (Admin)
      tags:
        - Admin
      parameters:
        - name: challenge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Danh sách bài nộp thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionList'
        '403':
          description: Không có quyền Admin.
  /admin/challenges/{challenge_id}/export-leaderboard:
    get:
      summary: Xuất Bảng xếp hạng ra file CSV (Admin)
      tags:
        - Admin
        - Leaderboard
      description: |
        Xuất toàn bộ BXH (cả public lẫn private) kèm thông tin Team, MSSV, Họ tên thành viên.
        File CSV encoding utf-8-sig (có BOM) để hiển thị đúng tiếng Việt trong Excel.
      parameters:
        - name: challenge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [public, private]
            default: private
          description: Loại BXH — "public" hoặc "private"
      responses:
        '200':
          description: File CSV
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'attachment; filename="leaderboard_{challenge_id}_{type}.csv"'
        '401':
          description: Chưa đăng nhập.
        '403':
          description: Không có quyền Admin.
        '404':
          description: Không tìm thấy bài thi.

  # ==========================================
  # Feature: Kernels / Solutions (UC: #27)
  # �?c t? TRU?C khi vi?t code � theo quy t?c 6.1 AGENTS.md
  # ==========================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT Token được lưu an toàn trong HttpOnly Cookie.
  parameters:
    PageParam:
      name: page
      in: query
      description: Vị trí trang hiện tại
      schema:
        type: integer
        default: 1
    SizeParam:
      name: size
      in: query
      description: Số lượng item trên một trang
      schema:
        type: integer
        default: 20
  schemas:
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 1500
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 75
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [STUDENT, ADMIN]
        is_active:
          type: boolean
          default: true
        student_id:
          type: string
          nullable: true
...
    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        public_score:
          type: number
          format: double
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCESS, FAILED]
        is_selected_for_private:
          type: boolean
        file_size_bytes:
          type: integer
          description: Kích thước file thực tế
        execution_time_ms:
          type: integer
          description: Thời gian Worker chấm bài (ms)
        error_message:
          type: string
          nullable: true
        submitted_at:
          type: string
          format: date-time
...